backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pypdf
sentence-transformers
faiss-cpu
openai
python-dotenv
numpy
